Extract factory method for view creation
