+++ lib/dashboard.js
@@ -3,16 +3,11 @@
 var _ = require("lodash");
 var blessed = require("blessed");
 
-var StreamView = require("./views/stream-view");
-var EventLoopView = require("./views/eventloop-view");
-var MemoryGaugeView = require("./views/memory-gauge-view");
-var MemoryGraphView = require("./views/memory-graph-view");
-var CpuView = require("./views/cpu-view");
 var HelpView = require("./views/help");
 var generateLayouts = require("./generate-layouts");
 var LogProvider = require("./providers/log-provider");
 var MetricsProvider = require("./providers/metrics-provider");
-var BaseView = require("./views/base-view");
+var views = require("./views");
 var GotoTimeView = require("./views/goto-time-view.js");
 
 var THROTTLE_TIMEOUT = 150;
@@ -127,14 +122,6 @@ Dashboard.prototype.onEvent = function (event) {
   }
 };
 
-var VIEW_MAP = {
-  log: StreamView,
-  cpu: CpuView,
-  memory: MemoryGaugeView,
-  memoryGraph: MemoryGraphView,
-  eventLoop: EventLoopView
-};
-
 Dashboard.prototype._showLayout = function (id) {
   if (this.currentLayout === id) {
     return;
@@ -146,14 +133,7 @@ Dashboard.prototype._showLayout = function (id) {
   this.views = [];
 
   _.each(this.layouts[id], function (layoutConfig) {
-    var View;
-
-    if (VIEW_MAP[layoutConfig.view.type]) {
-      View = VIEW_MAP[layoutConfig.view.type];
-    } else if (layoutConfig.view.module) {
-      // eslint-disable-next-line global-require
-      View = require(layoutConfig.view.module)(BaseView);
-    }
+    var View = views.getConstructor(layoutConfig.view);
 
     if (View) {
       if (this.settings[layoutConfig.view.type]) {
